+++ example/ios/Sharing Extension/ShareViewController.swift
@@ -15,8 +15,10 @@ class ShareViewController: SLComposeServiceViewController {
     // TODO: IMPORTANT: This should be your host app bundle identifier
     let hostAppBundleIdentifier = "com.kasem.sharing"
     let sharedKey = "ShareKey"
-    var sharedData: [String] = []
+    var sharedMedia: [SharedMediaFile] = []
+    var sharedText: [String] = []
     let imageContentType = kUTTypeImage as String
+    let videoContentType = kUTTypeMovie as String
     let textContentType = kUTTypeText as String
     let urlContentType = kUTTypeURL as String
     
@@ -37,6 +39,8 @@ class ShareViewController: SLComposeServiceViewController {
                         handleText(content: content, attachment: attachment, index: index)
                     } else if attachment.hasItemConformingToTypeIdentifier(urlContentType) {
                         handleUrl(content: content, attachment: attachment, index: index)
+                    } else if attachment.hasItemConformingToTypeIdentifier(videoContentType) {
+                        handleVideos(content: content, attachment: attachment, index: index)
                     }
                 }
             }
@@ -57,12 +61,12 @@ class ShareViewController: SLComposeServiceViewController {
             
             if error == nil, let item = data as? String, let this = self {
                 
-                this.sharedData.append(item)
+                this.sharedText.append(item)
                 
                 // If this is the last item, save imagesData in userDefaults and redirect to host app
                 if index == (content.attachments?.count)! - 1 {
                     let userDefaults = UserDefaults(suiteName: "group.\(this.hostAppBundleIdentifier)")
-                    userDefaults?.set(this.sharedData, forKey: this.sharedKey)
+                    userDefaults?.set(this.sharedText, forKey: this.sharedKey)
                     userDefaults?.synchronize()
                     this.extensionContext!.completeRequest(returningItems: [], completionHandler: nil)
                     this.redirectToHostApp(type: .text)
@@ -79,12 +83,12 @@ class ShareViewController: SLComposeServiceViewController {
             
             if error == nil, let item = data as? URL, let this = self {
                 
-                this.sharedData.append(item.absoluteString)
+                this.sharedText.append(item.absoluteString)
                 
                 // If this is the last item, save imagesData in userDefaults and redirect to host app
                 if index == (content.attachments?.count)! - 1 {
                     let userDefaults = UserDefaults(suiteName: "group.\(this.hostAppBundleIdentifier)")
-                    userDefaults?.set(this.sharedData, forKey: this.sharedKey)
+                    userDefaults?.set(this.sharedText, forKey: this.sharedKey)
                     userDefaults?.synchronize()
                     this.extensionContext!.completeRequest(returningItems: [], completionHandler: nil)
                     this.redirectToHostApp(type: .text)
@@ -96,45 +100,91 @@ class ShareViewController: SLComposeServiceViewController {
         }
     }
     
-    private func handleImages (content: NSExtensionItem, attachment: NSItemProvider, index: Int){
+    private func handleImages (content: NSExtensionItem, attachment: NSItemProvider, index: Int) {
         attachment.loadItem(forTypeIdentifier: imageContentType, options: nil) { [weak self] data, error in
             
             if error == nil, let url = data as? URL, let this = self {
                 
-                for component in url.path.components(separatedBy: "/") where component.contains("IMG_") {
-                    
-                    // photo: /var/mobile/Media/DCIM/101APPLE/IMG_1320.PNG
-                    // edited photo: /var/mobile/Media/PhotoData/Mutations/DCIM/101APPLE/IMG_1309/Adjustments/FullSizeRender.jpg
-                    
-                    // cut file's suffix if have, get file name like IMG_1309.
-                    let fileName = component.components(separatedBy: ".").first!
-                    if let asset = this.imageAssetDictionary[fileName] {
-                        this.sharedData.append( asset.localIdentifier)
-                    } else {
-                        // If we could not find the file then copy it
-                        let newPath = FileManager.default
-                            .containerURL(forSecurityApplicationGroupIdentifier: "group.\(this.hostAppBundleIdentifier)")!.appendingPathComponent(fileName)
-                        let copied = this.copyFile(at: url, to: newPath)
-                        if(copied) {
-                            this.sharedData.append(newPath.absoluteString)
-                        }
-                    }
-                    break
+                // Always copy
+                var fileExtension = url.lastPathComponent.components(separatedBy: ".")[safe: 1]
+                if fileExtension == nil {
+                    fileExtension = "png"
+                }
+                let newName = UUID().uuidString
+                let newPath = FileManager.default
+                    .containerURL(forSecurityApplicationGroupIdentifier: "group.\(this.hostAppBundleIdentifier)")!
+                    .appendingPathComponent("\(newName).\(fileExtension!)")
+                let copied = this.copyFile(at: url, to: newPath)
+                if(copied) {
+                    this.sharedMedia.append(SharedMediaFile(path: newPath.absoluteString, thumbnail: nil, duration: nil, type: .image))
                 }
                 
                 // If this is the last item, save imagesData in userDefaults and redirect to host app
                 if index == (content.attachments?.count)! - 1 {
-                    // TODO: IMPROTANT: This should be your host app bundle identiefier
-                    let hostAppBundleIdentiefier = "com.kasem.sharing"
-                    let userDefaults = UserDefaults(suiteName: "group.\(hostAppBundleIdentiefier)")
-                    userDefaults?.set(this.sharedData, forKey: this.sharedKey)
+                    let userDefaults = UserDefaults(suiteName: "group.\(this.hostAppBundleIdentifier)")
+                    userDefaults?.set(this.toData(data: this.sharedMedia), forKey: this.sharedKey)
                     userDefaults?.synchronize()
                     this.extensionContext!.completeRequest(returningItems: [], completionHandler: nil)
-                    this.redirectToHostApp(type: .image)
+                    this.redirectToHostApp(type: .media)
                 }
                 
             } else {
-                self?.dismissWithError()
+                print("GETTING ERROR")
+                let alert = UIAlertController(title: "Error", message: "Error loading image", preferredStyle: .alert)
+                
+                let action = UIAlertAction(title: "Error", style: .cancel) { _ in
+                    self?.dismiss(animated: true, completion: nil)
+                }
+                
+                alert.addAction(action)
+                self?.present(alert, animated: true, completion: nil)
+                self?.extensionContext!.completeRequest(returningItems: [], completionHandler: nil)
+            }
+        }
+    }
+    
+    private func handleVideos (content: NSExtensionItem, attachment: NSItemProvider, index: Int) {
+        attachment.loadItem(forTypeIdentifier: videoContentType, options: nil) { [weak self] data, error in
+            
+            if error == nil, let url = data as? URL, let this = self {
+                
+                // Always copy
+                var fileExtension = url.lastPathComponent.components(separatedBy: ".")[safe: 1]
+                if fileExtension == nil {
+                    fileExtension = "mp4"
+                }
+                let newName = UUID().uuidString
+                let newPath = FileManager.default
+                    .containerURL(forSecurityApplicationGroupIdentifier: "group.\(this.hostAppBundleIdentifier)")!
+                    .appendingPathComponent("\(newName).\(fileExtension!)")
+                let copied = this.copyFile(at: url, to: newPath)
+                if(copied) {
+                    guard let sharedFile = this.getSharedMediaFile(forVideo: newPath) else {
+                        return
+                    }
+                    this.sharedMedia.append(sharedFile)
+                }
+                
+                // If this is the last item, save imagesData in userDefaults and redirect to host app
+                if index == (content.attachments?.count)! - 1 {
+                    let userDefaults = UserDefaults(suiteName: "group.\(this.hostAppBundleIdentifier)")
+                    userDefaults?.set(this.toData(data: this.sharedMedia), forKey: this.sharedKey)
+                    userDefaults?.synchronize()
+                    this.extensionContext!.completeRequest(returningItems: [], completionHandler: nil)
+                    this.redirectToHostApp(type: .media)
+                }
+                
+            } else {
+                print("GETTING ERROR")
+                let alert = UIAlertController(title: "Error", message: "Error loading image", preferredStyle: .alert)
+                
+                let action = UIAlertAction(title: "Error", style: .cancel) { _ in
+                    self?.dismiss(animated: true, completion: nil)
+                }
+                
+                alert.addAction(action)
+                self?.present(alert, animated: true, completion: nil)
+                self?.extensionContext!.completeRequest(returningItems: [], completionHandler: nil)
             }
         }
     }
@@ -166,29 +216,9 @@ class ShareViewController: SLComposeServiceViewController {
     }
     
     enum RedirectType {
-        case image
+        case media
         case text
     }
-    
-    /// Key is the matched asset's original file name without suffix. E.g. IMG_193
-    private lazy var imageAssetDictionary: [String : PHAsset] = {
-        
-        let options = PHFetchOptions()
-        options.includeHiddenAssets = true
-        
-        let fetchResult = PHAsset.fetchAssets(with: options)
-        
-        var assetDictionary = [String : PHAsset]()
-        
-        for i in 0 ..< fetchResult.count {
-            let asset = fetchResult[i]
-            let fileName = asset.value(forKey: "filename") as! String
-            let fileNameWithoutSuffix = fileName.components(separatedBy: ".").first!
-            assetDictionary[fileNameWithoutSuffix] = asset
-        }
-        
-        return assetDictionary
-    }()
     
     func copyFile(at srcURL: URL, to dstURL: URL) -> Bool {
         do {
@@ -201,5 +231,70 @@ class ShareViewController: SLComposeServiceViewController {
             return false
         }
         return true
+    }
+    
+    private func getSharedMediaFile(forVideo: URL) -> SharedMediaFile? {
+        let asset = AVAsset(url: forVideo)
+        let duration = (CMTimeGetSeconds(asset.duration) * 1000).rounded()
+        let thumbnailPath = getThumbnailPath(for: forVideo)
+        
+        if FileManager.default.fileExists(atPath: thumbnailPath.path) {
+            return SharedMediaFile(path: forVideo.absoluteString, thumbnail: thumbnailPath.absoluteString, duration: duration, type: .video)
+        }
+        
+        var saved = false
+        let assetImgGenerate = AVAssetImageGenerator(asset: asset)
+        assetImgGenerate.appliesPreferredTrackTransform = true
+        //        let scale = UIScreen.main.scale
+        //        assetImgGenerate.maximumSize =  CGSize(width: 79 * scale, height: 79 * scale)
+        do {
+            let img = try assetImgGenerate.copyCGImage(at: CMTimeMakeWithSeconds(1.0, 600), actualTime: nil)
+            try UIImagePNGRepresentation(UIImage(cgImage: img))?.write(to: thumbnailPath)
+            saved = true
+        } catch {
+            saved = false
+        }
+        
+        return saved ? SharedMediaFile(path: forVideo.absoluteString, thumbnail: thumbnailPath.absoluteString, duration: duration, type: .video) : nil
+        
+    }
+    
+    private func getThumbnailPath(for url: URL) -> URL {
+        let fileName = Data(url.lastPathComponent.utf8).base64EncodedString().replacingOccurrences(of: "==", with: "")
+        let path = FileManager.default
+            .containerURL(forSecurityApplicationGroupIdentifier: "group.\(hostAppBundleIdentifier)")!
+            .appendingPathComponent("\(fileName).jpg")
+        return path
+    }
+    
+    class SharedMediaFile: Codable {
+        var path: String; // can be image, video or url path. It can also be text content
+        var thumbnail: String?; // video thumbnail
+        var duration: Double?; // video duration in milliseconds
+        var type: SharedMediaType;
+        
+        
+        init(path: String, thumbnail: String?, duration: Double?, type: SharedMediaType) {
+            self.path = path
+            self.thumbnail = thumbnail
+            self.duration = duration
+            self.type = type
+        }
+    }
+    
+    enum SharedMediaType: Int, Codable {
+        case image
+        case video
+    }
+    
+    func toData(data: [SharedMediaFile]) -> Data {
+        let encodedData = try? JSONEncoder().encode(data)
+        return encodedData!
+    }
+}
+
+extension Array {
+    subscript (safe index: UInt) -> Element? {
+        return Int(index) < count ? self[Int(index)] : nil
     }
 }