+++ example/lib/main.dart
@@ -12,7 +12,7 @@ class MyApp extends StatefulWidget {
 
 class _MyAppState extends State<MyApp> {
   StreamSubscription _intentDataStreamSubscription;
-  List<String> _sharedFiles;
+  List<SharedMediaFile> _sharedFiles;
   String _sharedText;
 
   @override
@@ -20,9 +20,10 @@ class _MyAppState extends State<MyApp> {
     super.initState();
 
     // For sharing images coming from outside the app while the app is in the memory
-    _intentDataStreamSubscription =
-        ReceiveSharingIntent.getImageStream().listen((List<String> value) {
+    _intentDataStreamSubscription = ReceiveSharingIntent.getMediaStream()
+        .listen((List<SharedMediaFile> value) {
       setState(() {
+        print("Shared:" + (_sharedFiles?.map((f) => f.path)?.join(",") ?? ""));
         _sharedFiles = value;
       });
     }, onError: (err) {
@@ -30,7 +31,7 @@ class _MyAppState extends State<MyApp> {
     });
 
     // For sharing images coming from outside the app while the app is closed
-    ReceiveSharingIntent.getInitialImage().then((List<String> value) {
+    ReceiveSharingIntent.getInitialMedia().then((List<SharedMediaFile> value) {
       setState(() {
         _sharedFiles = value;
       });
@@ -72,7 +73,7 @@ class _MyAppState extends State<MyApp> {
           child: Column(
             children: <Widget>[
               Text("Shared files:", style: textStyleBold),
-              Text(_sharedFiles?.join(",") ?? ""),
+              Text(_sharedFiles?.map((f) => f.path)?.join(",") ?? ""),
               SizedBox(height: 100),
               Text("Shared urls/text:", style: textStyleBold),
               Text(_sharedText ?? "")